pyproject.toml
[tool.poetry]
name = "qasys backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = "^2.8.2"
langchain = "^0.2.7"
chromadb = "^0.5.4"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.4"
langchain-community = "^0.2.7"
firebase-admin = "^6.5.0"
langchain-openai = "^0.1.16"
transformers = "^4.42.4"
langchain-huggingface = "^0.0.3"
ollama = "^0.2.1"
google-cloud-storage = "^2.17.0"
boto3 = "^1.34.144"
azure-storage-blob = "^12.20.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
